replace integer (uint32_t and int32_t) parameters by float - for peace of mind and in preparation for less arbitrary number ranges (better UX)

+++ src/driver_pwm.h
@@ -14,7 +14,7 @@ struct DriverPWM : public Driver {
 
   void update(Channel& c, uint32_t dt) {
     if(c.enabled) {
-      analogWrite(pin, max(0, min(1023,c.target)));
+      analogWrite(pin, max(0, min(1023,(int)c.target)));
       c.position = c.target; // no real feedback possible
     }
   };
